requirements.txt
python-binance
TA-Lib
pandas
numpy
websocket_client